requirements.txt
python-dotenv
langchain[openai]
fastapi
uvicorn
